+++ src/pages/Diary/index.tsx
@@ -1,8 +1,5 @@
 import { useState, useRef, useEffect } from "react";
 import styled from "styled-components";
-import { motion } from "framer-motion";
-import Flatpickr from "react-flatpickr";
-import "flatpickr/dist/themes/airbnb.css";
 import Sidebar from "../../components/Sidebar";
 import BGI from "../../asset/draft.png";
 import Button from "../../components/Button";
@@ -41,6 +38,7 @@ import MealSelector from "../../components/MealSelector";
 import { MealItem } from "../../types/mealTypes";
 import NutrientSelector from "../../components/FoodSelector/FoodSelector";
 import MoodSelector from "../../components/ＭoodSelector";
+import DatePicker from "../../components/DatePicker";
 
 type FoodItem = {
   id: string;
@@ -94,7 +92,6 @@ const Diary = () => {
   const [imageFile, setImageFile] = useState<File | null>(null);
   const [imagePreview, setImagePreview] = useState<string | null>(null);
   const foodSelectorRef = useRef<HTMLDivElement | null>(null);
-  const [isInline, setIsInline] = useState(true);
 
   const { addAlert, AlertMessage } = useAlert();
   const { state } = useLocation();
@@ -130,21 +127,6 @@ const Diary = () => {
       });
       annotation.show();
     }
-  }, []);
-  useEffect(() => {
-    const handleResize = () => {
-      if (window.innerWidth < 768) {
-        setIsInline(false);
-      } else {
-        setIsInline(true);
-      }
-    };
-
-    handleResize();
-
-    window.addEventListener("resize", handleResize);
-
-    return () => window.removeEventListener("resize", handleResize);
   }, []);
 
   const handleMouseEnter = (index: number) => {
@@ -303,25 +285,18 @@ const Diary = () => {
             />
           </FoodPickerContainer>
           <TimePickerWrapper>
-            <TimePickerContainer>
-              <TimePickerTitle
-                ref={(el) => (titleRefs.current[2] = el)}
-                onMouseEnter={() => handleMouseEnter(2)}
-                onMouseLeave={() => handleMouseLeave(2)}
-              >
-                Time
-                <RequiredMark />
-              </TimePickerTitle>
-              <StyledFlatpickr
-                value={selectedTime}
-                onChange={(date: Date[]) => setSelectedTime(date[0])}
-                options={{
-                  inline: isInline,
-                  enableTime: true,
-                  dateFormat: "Y-m-d H:i",
-                }}
-              />
-            </TimePickerContainer>
+            <TimePickerTitle
+              ref={(el) => (titleRefs.current[2] = el)}
+              onMouseEnter={() => handleMouseEnter(2)}
+              onMouseLeave={() => handleMouseLeave(2)}
+            >
+              Time
+              <RequiredMark />
+            </TimePickerTitle>
+            <DatePicker
+              initialTime={selectedTime.toISOString()}
+              onDateChange={setSelectedTime}
+            />
           </TimePickerWrapper>
         </FoodAndTimePickerWrapper>
 
@@ -460,60 +435,10 @@ const TimePickerWrapper = styled.div`
     width: 100%;
   }
 `;
-const TimePickerContainer = styled.div`
-  display: flex;
-  position: relative;
-  flex-direction: column;
-  width: 30%;
-  .flatpickr-calendar.inline {
-    top: 20px !important;
-  }
-
-  @media (max-width: 768px) {
-    margin: 24px 0;
-  }
-  @media (max-width: 480px) {
-    .flatpickr-calendar.inline {
-      right: 30px;
-      transform: scale(0.8);
-    }
-  }
-  @media (max-width: 360px) {
-    margin: 48px 0 0 0;
-    .flatpickr-calendar.inline {
-      top: 24px !important;
-      right: 46px;
-      transform: scale(0.7);
-    }
-  }
-  @media (max-width: 300px) {
-    margin: 48px 0 0 0;
-    .flatpickr-calendar.inline {
-      top: 12px !important;
-      right: 62px;
-      transform: scale(0.6);
-    }
-  }
-`;
 
 const TimePickerTitle = styled.h2`
-  width: 80px;
-`;
-
-const StyledFlatpickr = styled(Flatpickr)`
-  font-family: "KG Second Chances", sans-serif;
-  position: absolute;
-  left: 158px;
-  top: -8px;
   height: 50px;
-  width: 150px;
-  @media (max-width: 480px) {
-    left: 90px;
-  }
-  @media (max-width: 360px) {
-    top: 48px;
-    left: 0;
-  }
+  width: 80px;
 `;
 
 const MobileImageUploadContainer = styled.div`
